+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - remove legacy version annotation on docker-compose.yml
 
+### Changed
+- mark all phpunit provider as Static
+
 ## [1.2.1] - 2024-05-31
 ### Removed
 - drop tests support on Drupal <= 9.4